rag_project/requirements.txt
streamlit>=1.28.0
boto3>=1.28.0
faiss-cpu>=1.7.4
PyPDF2>=3.0.0
numpy>=1.24.0
python-dotenv>=1.0.0